@@ -1,36 +1,23 @@
 {
-  "name": "market-tracker",
+  "name": "my-app",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "plotly.js": "^1.58.5",
-    "query-string": "^7.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.2.0",
-    "react-multi-carousel": "^2.6.5",
-    "react-plotly.js": "^2.5.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.4.0",
+    "react-plotly.js": "^2.6.0",
     "react-router-dom": "^5.2.1",
-    "react-scripts": "^5.0.0",
-    "react-slick": "^0.28.1",
+    "react-scripts": "5.0.1",
+    "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1",
-    "socket.io-client": "^4.4.1",
-    "web-vitals": "^1.0.1"
+    "socket.io-client": "^4.4.1"
   },
   "scripts": {
-    "start": "react-scripts --max_old_space_size=4096 start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
   },
   "browserslist": {
     "production": [
