@@ -1,14 +1,12 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
-// import App from './App';
-import Routes from './Routes'
 
-ReactDOM.render(
-    <Routes />,
-  document.getElementById('root')
+import Routes from './Routes';
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <Routes />
+  </React.StrictMode>
 );
-
-
-
-
