@@ -1,5 +1,5 @@
+import { SERVER2 } from '../../config';
 import { io } from "socket.io-client";
-const URL = "http://localhost:3008";
-// const socket = io(URL);
+const URL = `${SERVER2}`;
 const socketLivePrice = io(URL, { autoConnect: false });
 export default socketLivePrice;
