@@ -1,5 +1,5 @@
+import { SERVER1 } from "../../config";
 import { io } from "socket.io-client";
-const URL = "http://localhost:3004";
-// const socket = io(URL);
+const URL = `${SERVER1}`;
 const socket = io(URL, { autoConnect: false });
 export default socket;
