@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 import { getWatchList,updateWatchList,deleteWatchList } from '../../adapters/watchlistApi';
-import { getToken, isAuthenticated } from '../../adapters/authApi';
+import { getToken, isAuthenticated } from '../../authentication/authApi';
 import { getStockHistory } from '../../adapters/userApi';
 
 const WatchListContext = (filterBtn,socket,socketLivePrice) => {
@@ -10,7 +10,6 @@ const WatchListContext = (filterBtn,socket,socketLivePrice) => {
 
     const [watchList,setWatchList]=useState(null);
     const [historicWatchList,setHistoricWatchList]=useState([]);
-
 
     const [initWatchlist,setInitWatchList]=useState(false);
     const [updateList,setUpdateList]=useState(false);
@@ -25,9 +24,7 @@ const WatchListContext = (filterBtn,socket,socketLivePrice) => {
         getWatchList(authInfo._id,token).then(watchList=>{
             setWatchList(watchList.watchList);
             setInitWatchList(true);
-            // alert('watchlist set');
         });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     },[]);
 
     useEffect(()=>{
@@ -35,9 +32,7 @@ const WatchListContext = (filterBtn,socket,socketLivePrice) => {
             socketLivePrice.disconnect();
             socket.disconnect();
             setInitWatchList(true);
-            // alert('WATCHLIST UPDATED SET'+JSON.stringify(watchList)+' AND initwatchlist is: '+initWatchlist);
         }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     },[updateList]);
 
     // init livePrices
@@ -54,16 +49,13 @@ const WatchListContext = (filterBtn,socket,socketLivePrice) => {
                 });
                 setLivePrices(newLivePrices);
                 setLivesPricesLoaded(true);
-                // alert('init live prices');
             }
         }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     },[initWatchlist])
 
     // init socket
     useEffect(()=>{
         if(livesPricesLoaded){
-            // alert('liveprices loaded');
             // socketlivePrice
             let id=authInfo._id;
             socketLivePrice.auth = { id };
@@ -84,21 +76,17 @@ const WatchListContext = (filterBtn,socket,socketLivePrice) => {
                 setLivePrices(newLivePrices);
                 socketLivePrice.emit('checkLivePrices',{});
             });
-            // alert('sockets set except updatewatchlistliveprice');
             setSocketsSet(true);
         }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     },[livesPricesLoaded]);
 
     useEffect(()=>{
         if(socketsSet){
             socketLivePrice.on('updateWatchlistLivePrice',({liveWatchList})=>{
                 if(JSON.stringify(liveWatchList)!==JSON.stringify(watchList)){
-                    // alert('comparing livewatchlist: '+JSON.stringify(liveWatchList)+' and watchlist: '+JSON.stringify(watchList));
                     socketLivePrice.disconnect();
                     socketLivePrice.removeAllListeners();
                     socket.disconnect();
-                    // alert('sockets all disconnected:livepricereached'+JSON.stringify(liveWatchList));
                     setInitWatchList(false);
                     setLivesPricesLoaded(false);
                     setSocketsSet(false);
@@ -107,17 +95,14 @@ const WatchListContext = (filterBtn,socket,socketLivePrice) => {
                 }
             });
         }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     },[socketsSet]);
     
-
     // init sockets
     useEffect(()=>{
         return ()=>{
             socket.disconnect();
             socketLivePrice.disconnect();
         };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     },[]);
 
     const filterUpdate = (filter)=>{
@@ -145,7 +130,6 @@ const WatchListContext = (filterBtn,socket,socketLivePrice) => {
             }));
         });
     };
-
     const cardDelete = (tickerSymbol)=>{
         deleteWatchList(authInfo._id,token,{symbol:tickerSymbol}).then(newWatchList=>{
             setWatchList(watchList.filter((stocks,i)=>{
@@ -156,17 +140,4 @@ const WatchListContext = (filterBtn,socket,socketLivePrice) => {
 
    return { watchList,historicWatchList,livePrices,livesPricesLoaded,filterUpdate,cardUpdate, cardDelete };
 };
-
 export { WatchListContext };
-
-
-
-
-
-
-
-
-
-
-
-
